Fix NMSAR 204 team members population
The NMSAR ICS 204 form does not have all of its team member fields labeled
in a rational order.  The team leader's name field is called "Text1" and
subsequent team member names are in fields called Name{n} where n ranges
from 2 through 8.  The resource name and skill fields are all identically
named with only the suffix 1-8 changing --- the 1 pertains to the team
leader.

As written originally, the code tried to stuff team members 2-8 into
fields 1-7, and the result was team members missing from the assignment
form.

This commit fixes the error, making sure that the team leader is put into
Text 1 (which was correct before) and not then overwritten by the next
team member (or made blank).

This was tested with real mission data, after observing that the form was
wrong with the original code.